% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about-me.R
\name{wigle_about_me}
\alias{wigle_about_me}
\title{Get WiGLE user object for the current logged-in user}
\usage{
wigle_about_me(api_key = wigle_api_key())
}
\arguments{
\item{api_key}{WiGLE API Key. See \code{\link[=wigle_api_key]{wigle_api_key()}}}
}
\value{
\code{list} or invisible JSON API response if unsuccessful and a 2xx.
}
\description{
Get WiGLE user object for the current logged-in user
}
