% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiglr-api-key.R
\name{wigle_api_key}
\alias{wigle_api_key}
\title{Get or set WIGLE_API_KEY value}
\usage{
wigle_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new WiGLE API key for the current environment?}
}
\value{
atomic character vector containing the WiGLE API key
}
\description{
The API wrapper functions in this package all rely on a WiGLE API
key residing in the environment variable \code{WIGLE_API_KEY}. The
easiest way to accomplish this is to set it in the \code{.Renviron} file in your
home directory.
}
\references{
\url{https://wigle.net/account}
}
