% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox-search.R
\name{wigle_bbox_search}
\alias{wigle_bbox_search}
\title{Get WiGLE named map of general statistics}
\usage{
wigle_bbox_search(lat1, lat2, lon1, lon2, only_mine = FALSE,
  search_after = NULL, api_key = wigle_api_key())
}
\arguments{
\item{lat1, lat2, lon1, lon2}{lower/upper lat/lon search area}

\item{only_mine}{<lgl> def: \code{FALSE}; Search only for points first discovered by the current user. Use any string to set, leave unset for general search. Can’t be used with COMMAPI auth, since these are points you have locally.}

\item{search_after}{if not \code{NULL} then the record to search after in paginated
responses which is returned in the results list as \code{search_after}.}

\item{api_key}{WiGLE API Key. See \code{\link[=wigle_api_key]{wigle_api_key()}}}
}
\value{
list or invisible JSON API response if unsuccessful
and a 2xx.
}
\description{
Get WiGLE named map of general statistics
}
\examples{
\dontrun{
wigle_bbox_search(43.2468, 43.2806, -70.9282, -70.8025)
}
}
