% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{wigle_country_stats}
\alias{wigle_country_stats}
\title{Get WiGLE statistics organized by country}
\usage{
wigle_country_stats(api_key = wigle_api_key())
}
\arguments{
\item{api_key}{WiGLE API Key. See \code{\link[=wigle_api_key]{wigle_api_key()}}}
}
\value{
data frame (tibble) or invisible JSON API response if unsuccessful
and a 2xx.
}
\description{
Get WiGLE statistics organized by country
}
