% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{wigle_region_stats}
\alias{wigle_region_stats}
\title{Get WiGLE statistics for a specified country, organized by region}
\usage{
wigle_region_stats(country = "US", api_key = wigle_api_key())
}
\arguments{
\item{country}{iso2c two-letter country code; defaults to \code{US}}

\item{api_key}{WiGLE API Key. See \code{\link[=wigle_api_key]{wigle_api_key()}}}
}
\value{
\code{list} of three data frames (tibbles) containing \code{region}, \code{encryption} and
\code{postal_code} stats or invisible JSON API response if unsuccessful
and a 2xx.
}
\description{
Get WiGLE statistics for a specified country, organized by region
}
