% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R, R/site.R
\name{wigle_site_stats}
\alias{wigle_site_stats}
\title{Get WiGLE named map of general statistics}
\usage{
wigle_site_stats(api_key = wigle_api_key())

wigle_site_stats(api_key = wigle_api_key())
}
\arguments{
\item{api_key}{WiGLE API Key. See \code{\link[=wigle_api_key]{wigle_api_key()}}}

\item{api_key}{WiGLE API Key. See \code{\link[=wigle_api_key]{wigle_api_key()}}}
}
\value{
list or invisible JSON API response if unsuccessful
and a 2xx.

data frame (tibble) or invisible JSON API response if unsuccessful
and a 2xx.
}
\description{
Get WiGLE named map of general statistics

Get WiGLE named map of site-level statistics
}
