% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiglr-package.R
\docType{package}
\name{wiglr}
\alias{wiglr}
\alias{wiglr-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/wiglr}
\item BugReports: \url{https://gitlab.com/hrbrmstr/wiglr/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
